// apps/web/lib/api/client/models/CriticalPathResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CriticalPathNode } from './CriticalPathNode';
export type CriticalPathResponse = {
    path?: Array<CriticalPathNode>;
    total_lag_hours?: (number | null);
};